+++ cm/cmind/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.19"
+__version__ = "0.7.19.1"
 
 from cmind.core import access
 from cmind.core import error